nenaprasno/requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
Flask-WTF
email-validator
flask-script
flask-migrate
flask-login
psycopg2
python-dotenv
python-docx
PyPDF2
pillow
pytesseract
requests